refactor: separate employee activity formatter
